refactor(settings): move branch-prefix setting to General
The branch-name prefix affects more than worktrees (it also feeds the
agent's branch-naming system prompt), so it shouldn't live on the
Worktrees page. Move it to General under a new "Git" group; the store
field and validation are unchanged.

Addresses review feedback on #2708.

Generated-By: PostHog Code
Task-Id: b378731a-ecbc-4db7-8da2-5414ba9a6fd3
